Gradient
Here is a version without the images.  At first, I did the 50%50% gradient but making it 50%51% smoothes the line a bit so it does not appear pixeled and now no matter how much you zoom in it will be nice and crisp. Hope this helps. 
Example 
linear-gradient( to left bottom, #E9E7DA  50%, #373F27 51%);